% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claymap3d.R
\name{claymap3d}
\alias{claymap3d}
\title{Make a 3D "Clay" Map}
\usage{
claymap3d(
  prepped_clay,
  shape_colour_var = NULL,
  colour_type = c("fill"),
  colour_alpha = 0.3,
  material = "clay",
  return_qmesh = TRUE,
  take_snapshot = TRUE,
  snapshot_filename = NULL
)
}
\arguments{
\item{prepped_clay}{The output of a "prep_clay()" function call.}

\item{shape_colour_var}{Variable name on which to base colour fill or border of shapes}

\item{colour_type}{Should the shapes' border or fill be coloured? One of 'fill','border'}

\item{colour_alpha}{Opacity of shape fill or border colour; defaults to 0.3 (30\%)}

\item{material}{The material to mold the 3d map from; one of 'clay', 'satellite'}

\item{return_qmesh}{Should this function return a 'qmesh' object?}

\item{take_snapshot}{Shall a 2D snapshot be taken of the 3d RGL widget? Defaults to TRUE}

\item{snapshot_filename}{What filename shall be used for the snap? Default is system date.}
}
\value{
A 3D map RGL widget to interact with; optionally, also a snapshot and a 'qmesh' object
}
\description{
Make a 3D "Clay" Map
}
\examples{
\dontrun
}
