% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_clay.R
\name{prep_clay}
\alias{prep_clay}
\title{Prep Clay for map-making.}
\usage{
prep_clay(shapes, map_detail = 6)
}
\arguments{
\item{shapes}{Either a single {sf} polygon object, or a list thereof, to map}

\item{map_detail}{The level of detail for the map; min is 1 and max is 14, defaults to 6}
}
\value{
A list of data objects needed to mold 2-d and 3-d clay maps
}
\description{
Prep Clay for map-making.
}
\examples{
\dontrun
}
