% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mold_claymap.R
\name{mold_claymap}
\alias{mold_claymap}
\title{Mold a Clay Map with Shapes on top}
\usage{
mold_claymap(
  prepped_clay,
  base_map_alpha = 0.5,
  shape_border_colour = "grey",
  shape_fill_vars = NULL,
  palettes = c("Spectral"),
  shape_fill_alphas = 0.1,
  background = TRUE
)
}
\arguments{
\item{prepped_clay}{This claymapR list object includes prepared data}

\item{base_map_alpha}{How transparent should the basemap be? Defaults to 0.5; higher values are more opaque.}

\item{shape_border_colour}{Which colour should delineate polygon borders?}

\item{shape_fill_vars}{Variable names on which to base colour/fill of points/shapes respectively}

\item{palettes}{Which palettes should be used for shapes' colour or fill?}

\item{shape_fill_alphas}{How transparent should the shapes' colour(s)/fill(s) be?}

\item{background}{Should an 'ocean' background be added?}
}
\value{
A ggplot2 plot of your shape(s) with a basemap below.
}
\description{
Mold a Clay Map with Shapes on top
}
\examples{
\dontrun
}
