% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mold_claymap.R
\name{mold_claymap}
\alias{mold_claymap}
\title{Mold a Clay Map with Shapes on top}
\usage{
mold_claymap(
  prepped_clay,
  no_legend = TRUE,
  base_map_alpha = 0.5,
  shape_border_colour = "grey",
  shape_fill_var = NULL,
  shape_fill_alpha = 0.1,
  background = TRUE,
  verbose = F
)
}
\arguments{
\item{no_legend}{Should the fill colour legend be dropped from the final map?}

\item{base_map_alpha}{How transparent should the basemap be? Defaults to 0.5; higher values are more opaque.}

\item{shape_border_colour}{What colour should delineate the shapes' borders?}

\item{shape_fill_var}{Variable name on which to base colour fill of shapes}

\item{shape_fill_alpha}{How transparent should the shapes' fill colour be?}

\item{background}{Should an 'ocean' background be added?}

\item{verbose}{Return copious feedback?}

\item{shapes}{Shapes to map}

\item{map_detail}{If base_type is 'clay', a value from 1 (least detailed) to 14 (most detailed); default is 6}

\item{elevation_map}{A {terra} raster of elevation for the area of interest; default is NULL and will calculate}
}
\value{
A ggplot2 plot of your shape(s) with a basemap below.
}
\description{
Mold a Clay Map with Shapes on top
}
\examples{
\dontrun
}
